feat: basic server
Start a GenServer. Check if there's a successful connection to Telegram first,
and then we will proceed with doing the other cooler things in the subsequent
commits.

Add this server as a children of our main application, so it initialises on
startup automatically. Plus we get all the fault-tolerance perks.