fix: handle non-message updates from Telegram
Editing a message sent to the bot triggers an error, since the bot assumes that
all message must have a "message" key.

A quick glance at the documentation here reveals that this is not the case:

  https://core.telegram.org/bots/api#update

"message" is marked as optional.

The current mechanism of long polling does not allow us to skip messages when
`parse_update_msg/1` fails. Alleviate this situation by ignoring all other types
of updates. "update_id" is not an optional field according to the documentation.