refactor: move Telegram-specific code from server to client
Move code related to processing response from Telegram API calls into the
Telegram client.

Pass per-update processing function as a callback into the Telegram client. In
conjunction with changing `Enum.map(updates, &parse_update_msg/1)` to
`Stream.map`, this allows us to parse the message and perform the callback for
each item in the updates list in a single pass. This may not improve the
performance by an amazing margin (or at all), but it feels like the more elegant
approach.